Removed unused URLConvertible protocol and replaced with the ExpressibleByStringLiteral conformance
